+++ README.md
@@ -1,5 +1,12 @@
 # Docker image for postgis with osm2pgsql & shp2pgsql
 
+データベースシステム講義資料
+version 0.0.1
+authors: N. Tsutsumida
+
+Copyright (c) 2023 Narumasa Tsutsumida
+Released under the MIT license
+https://opensource.org/licenses/mit-license.php
 
 ## Prerequisites
 - DockerとVScodeのインストール
