--- work/python/sample/sample_pd_py/sample_pd_3.py
@@ -1,25 +1,0 @@
-import os
-from sqlalchemy import create_engine
-import pandas as pd
-pd.set_option('display.max_columns', 100)
-
-def query_pandas(db):
-    DATABASE_URL='postgresql://postgres:postgres@postgis_container:5432/{}'.format(db)
-    conn = create_engine(DATABASE_URL)
-
-    #埼玉県内の市町村の面積はいくつ？
-    sql = "select name_2, st_area(geom::geography)/1000000 as area_km2 from adm2 where name_1='Saitama' order by area_km2 desc;"
-    print(sql)
-
-    df = pd.read_sql(sql=sql, con=conn)
-
-    return df
-
-
-def main():
-
-    out = query_pandas('gisdb') #specify db name
-    print(out)
-
-if __name__ == '__main__':
-    main()
