--- work/python/sample/sample_pd_py/sample_pd_5.py
@@ -1,30 +1,0 @@
-import os
-from sqlalchemy import create_engine
-import pandas as pd
-pd.set_option('display.max_columns', 100)
-
-def query_pandas(db):
-    DATABASE_URL='postgresql://postgres:postgres@postgis_container:5432/{}'.format(db)
-    conn = create_engine(DATABASE_URL)
-
-    #県ごとに一番小さい面積を有する市町村を調べる
-    sql = "select name_1, name_2, st_area(geom::geography)/1000000 as area_km2 \
-            from adm2 \
-            where st_area(geom::geography)/1000000 in \
-            (select min(st_area(geom::geography)/1000000) from adm2 group by name_1) \
-            order by area_km2 asc;"
-
-    print(sql)
-
-    df = pd.read_sql(sql=sql, con=conn)
-
-    return df
-
-
-def main():
-
-    out = query_pandas('gisdb') #specify db name
-    print(out)
-
-if __name__ == '__main__':
-    main()
